Fix visibleSets adding ALL tab back to UI
The visibleSets memo was re-adding the ALL option even though it was
removed from constants.js. This caused the ALL and ALT tabs to appear
in the UI but be non-functional.

Changed visibleSets logic to simply show SETS or filtered sets based
on availableSets, without re-adding ALL.

Note: ALT (A Lawless Time) is a legitimate set, not a bug.